+++ services/CHANGELOG.md
@@ -19,7 +19,10 @@
 - preprocess-extract:
   - [Improvement] JSON flow body refactor (preprocess and indexation)
 - preprocess-ocr:
-  - [New] Changes to support 'llm-ocr' (minimum genai-sdk-services version==0.5.0 )
+  - [New] Changes to support 'llm-ocr' (minimum genai-sdk-services version==0.5.2)
+    - New required environment variables:
+      - Queue: 'Q_GENAI_LLMQUEUE_INPUT', 'Q_GENAI_LLMQUEUE_OUTPUT'
+      - API: 'URL_LLM'
   - [Improvement] JSON flow body refactor (preprocess and indexation)
   - [Improvement] Resize image method generalization and moved to 'utils.py'
 - genai-infoindexing:
@@ -31,6 +34,9 @@
     - To support this feature, all images are now sent to the model in base64 format (URL in input still working)
   - [Improvement] If not 'x-limits' in headers, call apigw to get current limits (mainly for queue case)
   - [Improvement] Add support for vision in 'MOUNT_KEY' feature
+- genai-inforetrieval
+  - [Improvement] Adapt input JSON for coherence purposes with json body refactor
+
 
 
 ## v2.2.0 (2024-12-19)